Move the modules around, prepare for separation of indexer, enricher and server
