+++ addons/input_module/localfiles/enricher/__init__.py
@@ -1,0 +1,17 @@
+# Export main enricher functionality
+from .enricher import (
+    start_enricher,
+    stop_enricher,
+    trigger_enrichment,
+    EnrichmentStatus,
+)
+
+# Export database class
+from .enricher_db import EnricherDb
+
+# Export enricher plugins
+from .enricher_plugin import EnricherPlugin
+from .acoustid_plugin import AcoustIdPlugin
+from .musicbrainz_plugin import MusicBrainzPlugin
+from .wikidata_plugin import WikidataPlugin
+from .deezer_plugin import DeezerPlugin
