+++ addons/input_module/localfiles/indexer/__init__.py
@@ -1,0 +1,12 @@
+# Export main indexer functionality
+from .indexer import (
+    start_indexer,
+    stop_indexer,
+    trigger_scan,
+    register_enricher_callback,
+    start_file_watcher,
+    stop_file_watcher,
+)
+
+# Export database class
+from .indexer_db import IndexerDb
